Provide constraint tuples up to size 64
GHC 9.2+ bumps the value of `maxTupleSize` to `64`, so we follow suit here in
`constraint-tuples`. Note that GHC 8.0 through 9.0 impose a maximum tuple size
of 62, so we can only define 63- and 64- tuples on GHC 9.2 or later.

In an effort to make the output of the generator script the same across all
versions of GHC, we now hard-code the value of `maxTupleSize` rather than
importing it from `GHC.Exts` (where its value can change between GHC versions).